Voice_ai/requirements.txt
aiofiles                     23.2.1
altair                       5.3.0
annotated-types              0.6.0
anyio                        4.3.0
argon2-cffi                  23.1.0
argon2-cffi-bindings         21.2.0
arrow                        1.3.0
asttokens                    2.4.1
async-lru                    2.0.4
attrs                        23.2.0
audioread                    3.0.1
Babel                        2.14.0
beautifulsoup4               4.12.3
bleach                       6.1.0
cachetools                   5.3.3
certifi                      2024.2.2
cffi                         1.16.0
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
comm                         0.2.2
contourpy                    1.2.0
cycler                       0.12.1
debugpy                      1.8.1
decorator                    5.1.1
defusedxml                   0.7.1
exceptiongroup               1.2.0
executing                    2.0.1
fastapi                      0.110.0
fastjsonschema               2.19.1
ffmpy                        0.3.2
filelock                     3.13.1
fonttools                    4.50.0
fqdn                         1.5.1
fsspec                       2024.3.1
gitdb                        4.0.11
GitPython                    3.1.42
google-ai-generativelanguage 0.4.0
google-api-core              2.18.0
google-auth                  2.29.0
google-generativeai          0.4.1
googleapis-common-protos     1.63.0
gradio                       4.24.0
gradio_client                0.14.0
grpcio                       1.62.1
grpcio-status                1.62.1
h11                          0.14.0
httpcore                     1.0.4
httpx                        0.27.0
huggingface-hub              0.22.2
idna                         3.6
importlib_resources          6.4.0
ipykernel                    6.29.3
ipython                      8.22.2
isoduration                  20.11.0
jedi                         0.19.1
Jinja2                       3.1.3
joblib                       1.3.2
json5                        0.9.24
jsonpointer                  2.4
jsonschema                   4.21.1
jsonschema-specifications    2023.12.1
jupyter_client               8.6.1
jupyter_core                 5.7.2
jupyter-events               0.10.0
jupyter-lsp                  2.2.4
jupyter_server               2.13.0
jupyter-server-mathjax       0.2.6
jupyter_server_terminals     0.5.3
jupyterlab                   4.1.5
jupyterlab_git               0.50.0
jupyterlab_pygments          0.3.0
jupyterlab_server            2.25.4
kiwisolver                   1.4.5
lazy_loader                  0.3
librosa                      0.10.1
llvmlite                     0.42.0
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
matplotlib                   3.8.3
matplotlib-inline            0.1.6
mdurl                        0.1.2
mistune                      3.0.2
mpmath                       1.3.0
msgpack                      1.0.8
nbclient                     0.10.0
nbconvert                    7.16.2
nbdime                       4.0.1
nbformat                     5.10.3
nest-asyncio                 1.6.0
networkx                     3.2.1
notebook_shim                0.2.4
numba                        0.59.1
numpy                        1.26.4
nvidia-cublas-cu12           12.1.3.1
nvidia-cuda-cupti-cu12       12.1.105
nvidia-cuda-nvrtc-cu12       12.1.105
nvidia-cuda-runtime-cu12     12.1.105
nvidia-cudnn-cu12            8.9.2.26
nvidia-cufft-cu12            11.0.2.54
nvidia-curand-cu12           10.3.2.106
nvidia-cusolver-cu12         11.4.5.107
nvidia-cusparse-cu12         12.1.0.106
nvidia-nccl-cu12             2.19.3
nvidia-nvjitlink-cu12        12.4.99
nvidia-nvtx-cu12             12.1.105
orjson                       3.10.0
overrides                    7.7.0
packaging                    24.0
pandas                       2.2.1
pandocfilters                1.5.1
parso                        0.8.3
pexpect                      4.9.0
pillow                       10.2.0
pip                          24.0
platformdirs                 4.2.0
plotly                       5.20.0
pooch                        1.8.1
prometheus_client            0.20.0
prompt-toolkit               3.0.43
proto-plus                   1.23.0
protobuf                     4.25.3
psutil                       5.9.8
ptyprocess                   0.7.0
pure-eval                    0.2.2
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pycparser                    2.21
pydantic                     2.6.4
pydantic_core                2.16.3
pydub                        0.25.1
Pygments                     2.17.2
pyparsing                    3.1.2
python-dateutil              2.9.0.post0
python-json-logger           2.0.7
python-multipart             0.0.9
pytz                         2024.1
PyYAML                       6.0.1
pyzmq                        25.1.2
referencing                  0.34.0
regex                        2023.12.25
requests                     2.31.0
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rich                         13.7.1
rpds-py                      0.18.0
rsa                          4.9
ruff                         0.3.4
safetensors                  0.4.2
scikit-learn                 1.4.1.post1
scipy                        1.12.0
seaborn                      0.13.2
semantic-version             2.10.0
Send2Trash                   1.8.2
setuptools                   68.2.2
shellingham                  1.5.4
six                          1.16.0
smmap                        5.0.1
sniffio                      1.3.1
soundfile                    0.12.1
soupsieve                    2.5
soxr                         0.3.7
SpeechRecognition            3.10.3
stack-data                   0.6.3
starlette                    0.36.3
sympy                        1.12
tenacity                     8.2.3
terminado                    0.18.1
threadpoolctl                3.4.0
tinycss2                     1.2.1
tokenizers                   0.15.2
tomli                        2.0.1
tomlkit                      0.12.0
toolz                        0.12.1
torch                        2.2.2
torchaudio                   2.2.2
tornado                      6.4
tqdm                         4.66.2
traitlets                    5.14.2
transformers                 4.39.2
triton                       2.2.0
typer                        0.12.0
typer-cli                    0.12.0
typer-slim                   0.12.0
types-python-dateutil        2.9.0.20240316
typing_extensions            4.10.0
tzdata                       2024.1
uri-template                 1.3.0
urllib3                      2.0.7
uvicorn                      0.29.0
wcwidth                      0.2.13
webcolors                    1.13
webencodings                 0.5.1
websocket-client             1.7.0
websockets                   11.0.3